+++ test/Audit.Integration.AspNetCore/WebApiTests.cs
@@ -5,7 +5,6 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Audit.Core;
-using Audit.Integration.AspNetCore.Controllers;
 using Audit.WebApi;
 using Newtonsoft.Json;
 using NUnit.Framework;
@@ -25,6 +24,30 @@ namespace Audit.Integration.AspNetCore
             var c = new HttpClient();
             var s = await c.GetStringAsync($"http://localhost:{_port}/api/values");
             Assert.AreEqual("[\"value1\",\"value2\"]", s);
+        }
+
+        public async Task Test_WebApi_GlobalFilter_Async()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicAsyncProvider(_ => _.OnInsert(async ev =>
+                {
+                    await Task.Delay(1);
+                    insertEvs.Add(ev);
+                    return Guid.NewGuid();
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var c = new HttpClient();
+            var s = await c.PostAsync($"http://localhost:{_port}/api/values/GlobalAudit", new StringContent("{\"value\": \"def\"}", Encoding.UTF8, "application/json"));
+
+
+            Assert.AreEqual(HttpStatusCode.OK, s.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            Assert.AreEqual("{\"value\": \"def\"}", insertEvs[0].GetWebApiAuditAction().RequestBody.Value);
+            Assert.AreEqual("def", insertEvs[0].GetWebApiAuditAction().ResponseBody.Value);
+            Assert.AreEqual(200, insertEvs[0].GetWebApiAuditAction().ResponseStatusCode);
+            Assert.AreEqual("POST.Values.GlobalAudit", insertEvs[0].EventType);
         }
 
         public async Task Test_WebApi_HappyPath_Async()